pyproject.toml
[project]
name = "ai-gym-timetable-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "certifi>=2024.0.0",
]

[tool.setuptools]
py-modules = ["GymScheduleExtractor", "OcrEngine", "Constants"]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "mypy"
]
test = [
    "pytest",
    "pytest-cov"
]